# docs/design/api.yaml
openapi: 3.0.3
info:
  title: 勤怠管理システム API
  version: 1.0.0
  description: 中小企業向け勤怠管理システムの REST API
servers:
  - url: http://localhost:8080/api
    description: ローカル開発
paths:
...
  /auth/login:
    post:
      tags: [認証]
      summary: ログイン
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 認証失敗
  /auth/me:
    get:
      tags: [認証]
      summary: ログインユーザー情報取得
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'

  # === 打刻 ===
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        employee:
          $ref: '#/components/schemas/EmployeeResponse'
    EmployeeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [EMPLOYEE, ADMIN]
...